requirements.txt
oauth2client<4.0.0
google-api-python-client>=2.0.0
requests>=2.0.0
Pillow>=9.0.0
torch>=2.0.0
numpy>=1.0.0